Renamed student enrollment handler from 'dummyQuery' to 'student_enrollment'
